move endpoints under routes
