+++ v0/routes/register/routes.js
@@ -1,4 +1,4 @@
-const { json } = require('../middleware')
+const { json } = require('../../middleware')
 const registerUser = require('./registerUser')
 
 module.exports = function(app) {
